Movement works in all directions

+++ concave/constants.h
@@ -86,4 +86,4 @@ const float MIN_FRAME_RATE = 10.0f;             // the minimum frame rate
 const float MIN_FRAME_TIME = 1.0f / FRAME_RATE;   // minimum desired time for 1 frame
 const float MAX_FRAME_TIME = 1.0f / MIN_FRAME_RATE; // maximum time used in calculations
 
-const float PI = 3.141592653589793238463;
+const double PI = 3.141592653589793238463;
